+++ amdsmi_cli/amdsmi_parser.py
@@ -607,6 +607,10 @@ class AMDSMIParser(argparse.ArgumentParser):
         soc_pstate_help = "The available soc pstate policy"
         xgmi_plpd_help = "The available XGMI per-link power down policy"
         process_isolation_help = "The process isolation status"
+        clk_options = self.helpers.get_clock_types()[0]
+        clk_options.remove('PCIE')
+        clk_option_str = ", ".join(clk_options) + ", ALL"
+        clock_help = f"Show one or more valid clock frequency levels. Available options:\n\t{clk_option_str}"
 
         # Options arguments help text for Hypervisors and Baremetal
         ras_help = "Displays RAS features information"
@@ -642,6 +646,7 @@ class AMDSMIParser(argparse.ArgumentParser):
             static_parser.add_argument('-B', '--board', action='store_true', required=False, help=board_help)
             static_parser.add_argument('-R', '--process-isolation', action='store_true', required=False, help=process_isolation_help)
             static_parser.add_argument('-r', '--ras', action='store_true', required=False, help=ras_help)
+            static_parser.add_argument('-C', '--clock', default=False, nargs='*', type=str, required=False, help=clock_help)
 
             # Options to display on Hypervisors and Baremetal
             if self.helpers.is_hypervisor() or self.helpers.is_baremetal():